+++ protech-commerce/includes/class-messaging-tab.php
@@ -135,6 +135,7 @@ class MessagingTab {
 			'automations' => __( 'Emails', 'protech-wholesale' ),
 			'compose'     => __( 'Compose', 'protech-wholesale' ),
 			'templates'   => __( 'Email templates', 'protech-wholesale' ),
+			'contacts'    => __( 'Contacts', 'protech-wholesale' ),
 			'log'         => __( 'Log', 'protech-wholesale' ),
 			'compliance'  => __( 'Compliance', 'protech-wholesale' ),
 			'settings'    => __( 'Settings', 'protech-wholesale' ),
@@ -194,6 +195,9 @@ class MessagingTab {
 				break;
 			case 'templates':
 				EmailComposer::render();
+				break;
+			case 'contacts':
+				ContactsScreen::render();
 				break;
 			case 'log':
 				LogScreen::render();